+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## unreleased
 ### Changed
 * Improved error messages when importing or exporting a nud model without the required vbn file.
+* Improved generated material nodes to include basic assignments for color and normal textures.
 
 ## 0.1.0 - 2025-08-07
 First public release!